+++ src/lib/kombi/TaskHolder/Dispatcher/Renderfarm/Job/__init__.py
@@ -1,3 +1,3 @@
-from .RenderfarmJob import RenderfarmJob, RenderfarmJobIdError
+from .Job import Job, JobIdError
 from .ExpandedJob import ExpandedJob
 from .CollapsedJob import CollapsedJob
