--- src/lib/kombi/Template/Procedure/__init__.py
@@ -1,7 +1,0 @@
-from . import Math
-from . import ImageSequence
-from . import Text
-from . import Path
-from . import System
-from . import Version
-from . import Datetime
