+++ test/TaskHolder/Dispatcher/Local/__init__.py
@@ -1,1 +1,1 @@
-from .LocalTest import LocalTest
+from .LocalDispatcherTest import LocalDispatcherTest
