--- test/Template/Procedure/__init__.py
@@ -1,7 +1,0 @@
-from .DatetimeTest import DatetimeTest
-from .ImageSequenceTest import ImageSequenceTest
-from .MathTest import MathTest
-from .PathTest import PathTest
-from .SystemTest import SystemTest
-from .TextTest import TextTest
-from .VersionTest import VersionTest
