+++ test/Template/__init__.py
@@ -1,2 +1,2 @@
 from .TemplateTest import TemplateTest
-from . import Procedure
+from . import procedures
