+++ test/Template/procedures/DatetimeProceduresTest.py
@@ -2,7 +2,7 @@ import unittest
 from ...BaseTestCase import BaseTestCase
 from kombi.Template import Template
 
-class DatetimeTest(BaseTestCase):
+class DatetimeProceduresTest(BaseTestCase):
     """Test Datetime template procedures."""
 
     def testYear(self):
